Initiate building email invitation button
